Finalizing graph's depth first traversal

+++ src/Vector2.hpp
@@ -5,7 +5,7 @@
 struct Vector2 {
 	float x{}, y{};
 
-	Vector2(int x = 0, int y = 0)
+	Vector2(float x = 0.0f, float y = 0.0f)
 	{
 		this->x = x;
 		this->y = y;
@@ -13,8 +13,6 @@ struct Vector2 {
 
   ~Vector2() = default;
 
-	// QUESTION
-	// Why does it have to be declared friend?
 	friend bool operator== (const Vector2& left, const Vector2& right) {
 		if (left.x == right.x || left.y == right.y)
 				return true;
@@ -27,11 +25,10 @@ struct std::hash<Vector2>
 {
 	std::size_t operator()(const Vector2& v) const noexcept
 	{
-		// QUESTION
-		// To be honest, I do not really know what I am doing, just following cppreference.
+		// To be honest, I do not really know what I am doing, just copied cppreference...
 		std::size_t h1 = std::hash<int>{}(v.x);
 		std::size_t h2 = std::hash<int>{}(v.y);
-		return h1 ^ (h2 << 1);
+		return h1 ^ (h2 << 1); // Bitwise operators. Optimizing store bools in ints and stuff
 	}
 
 };
